Working `GetD3DTexture` and `GetLocalCubemap` implemented

+++ modules/gelly-binary-module/src/source/CTexture.h
@@ -2,7 +2,7 @@
 #define GELLY_CTEXTURE_H
 
 typedef void **CTexture;
-typedef int ShaderAPITextureHandle_t;
+typedef unsigned long long ShaderAPITextureHandle_t;
 
 ShaderAPITextureHandle_t GetCTextureHandle(CTexture *texture);
 const char *GetCTextureName(CTexture *texture);
